app/requirements.txt
scikit-learn==1.2.2
tensorflow==2.12.0
keras==2.12.0
h5py==3.9.0
Pillow==9.5.0
pandas==2.0.3
numpy==1.23.5
matplotlib
seaborn
time
